do not put different heuristic types (message, issues, changes) in different files, instead put them in one with the name which corresponds to the task heuristics are written for (e.g. bugginess)
