+++ dvc.yaml
@@ -7,10 +7,10 @@ stages:
     - data/train:
         cache: false
   apply_heuristics_message:
-    cmd: python bohr/pipeline/apply_heuristics.py message --n-parallel=1 --commits-file data/train/bug_sample.csv --issues-file data/train/bug_sample_issues.csv
+    cmd: python bohr/pipeline/apply_heuristics.py bugginess --n-parallel=1 --commits-file data/train/bug_sample.csv --issues-file data/train/bug_sample_issues.csv --changes-file data/train/bug_sample_files.csv
     deps:
     - bohr/heuristics/keywords
-    - bohr/heuristics/message.py
+    - bohr/heuristics/bugginess.py
     - bohr/pipeline/apply_heuristics.py
     - data/train
     - data/test
@@ -30,59 +30,8 @@ stages:
     metrics:
     - metrics/message/heuristic_metrics.json:
         cache: false
-  apply_heuristics_message_issues:
-    cmd: python bohr/pipeline/apply_heuristics.py message issues --n-parallel=1 --commits-file data/train/bug_sample.csv --issues-file data/train/bug_sample_issues.csv
-    deps:
-      - bohr/heuristics/keywords
-      - bohr/heuristics/message.py
-      - bohr/heuristics/issues.py
-      - bohr/pipeline/apply_heuristics.py
-      - data/train
-      - data/test
-    outs:
-      - generated/message_issues/heuristic_matrix_train.pkl
-      - generated/message_issues/heuristic_matrix_herzig.pkl
-      - generated/message_issues/heuristic_matrix_berger.pkl
-      - generated/message_issues/heuristic_matrix_1151_commits.pkl
-      - generated/message_issues/analysis_train.csv:
-          cache: false
-      - generated/message_issues/analysis_herzig.csv:
-          cache: false
-      - generated/message_issues/analysis_berger.csv:
-          cache: false
-      - generated/message_issues/analysis_1151_commits.csv:
-          cache: false
-    metrics:
-      - metrics/message_issues/heuristic_metrics.json:
-          cache: false
-  apply_heuristics_message_issues_changes:
-    cmd: python bohr/pipeline/apply_heuristics.py message issues changes --n-parallel=1 --commits-file data/train/bug_sample.csv --issues-file data/train/bug_sample_issues.csv --changes-file data/train/bug_sample_files.csv
-    deps:
-      - bohr/heuristics/keywords
-      - bohr/heuristics/message.py
-      - bohr/heuristics/issues.py
-      - bohr/heuristics/changes.py
-      - bohr/pipeline/apply_heuristics.py
-      - data/train
-      - data/test
-    outs:
-      - generated/message_issues_changes/heuristic_matrix_train.pkl
-      - generated/message_issues_changes/heuristic_matrix_herzig.pkl
-      - generated/message_issues_changes/heuristic_matrix_berger.pkl
-      - generated/message_issues_changes/heuristic_matrix_1151_commits.pkl
-      - generated/message_issues_changes/analysis_train.csv:
-          cache: false
-      - generated/message_issues_changes/analysis_herzig.csv:
-          cache: false
-      - generated/message_issues_changes/analysis_berger.csv:
-          cache: false
-      - generated/message_issues_changes/analysis_1151_commits.csv:
-          cache: false
-    metrics:
-      - metrics/message_issues_changes/heuristic_metrics.json:
-          cache: false
   train_label_model_message:
-    cmd: python bohr/pipeline/train_label_model.py message
+    cmd: python bohr/pipeline/train_label_model.py bugginess
     deps:
     - bohr/pipeline/train_label_model.py
     - generated/message/heuristic_matrix_train.pkl
@@ -96,42 +45,11 @@ stages:
     metrics:
     - metrics/message/label_model_metrics.json:
         cache: false
-  train_label_model_message_issues:
-    cmd: python bohr/pipeline/train_label_model.py message issues
-    deps:
-      - bohr/pipeline/train_label_model.py
-      - generated/message_issues/heuristic_matrix_train.pkl
-      - generated/message_issues/heuristic_matrix_herzig.pkl
-      - generated/message_issues/heuristic_matrix_berger.pkl
-      - generated/message_issues/heuristic_matrix_1151_commits.pkl
-      - data/train
-      - data/test
-    outs:
-      - generated/message_issues/label_model.pkl
-    metrics:
-      - metrics/message_issues/label_model_metrics.json:
-          cache: false
   label_dataset_message_issues:
-    cmd: python bohr/pipeline/label_dataset.py message issues --commits-file data/train/bug_sample.csv -o labeled_with_model2.csv --debug
+    cmd: python bohr/pipeline/label_dataset.py bugginess --commits-file data/train/bug_sample.csv -o labeled_with_model2.csv --debug
     deps:
       - bohr/pipeline/label_dataset.py
       - generated/message_issues/heuristic_matrix_train.pkl
       - data/train
     outs:
       - labeled_with_model2.csv
-  train_label_model_message_issues_changes:
-    cmd: python bohr/pipeline/train_label_model.py message issues changes
-    deps:
-      - bohr/pipeline/train_label_model.py
-      - generated/message_issues_changes/heuristic_matrix_train.pkl
-      - generated/message_issues_changes/heuristic_matrix_herzig.pkl
-      - generated/message_issues_changes/heuristic_matrix_berger.pkl
-      - generated/message_issues_changes/heuristic_matrix_1151_commits.pkl
-      - data/train
-      - data/test
-    outs:
-      - generated/message_issues_changes/label_model.pkl
-    metrics:
-      - metrics/message_issues_changes/label_model_metrics.json:
-          cache: false
-
